\name{techila-package}
\alias{techila-package}
\alias{techila}
\docType{package}
\title{
Techila R package
}
\description{
This package is used to command Techila from within GNU R.
}
\details{
\tabular{ll}{
Package: \tab techila\cr
Type: \tab Package\cr
Version: \tab 3.6\cr
Date: \tab 2017-11-01\cr
License: \tab Proprietary\cr
LazyLoad: \tab yes\cr
}
To use the Techila from within R, use peach or cloudfor function.
}
\author{
Author: Techila Technologies Ltd.
Contact: Techila Support <support@techilatechnologies.com>
Maintainer: Kari Koskinen <kari.koskinen@techilatechnologies.com>
}
\keyword{ techila }
